fix mutable state example
Introduces new mutable_state.rs with code moved from state.rs.
Updates application.md to reference this file.
Removes un-necessary underscore from "index" function in
mutable-state.rs.
