+++ examples/application/src/state.rs
@@ -14,39 +14,6 @@ async fn index(data: web::Data<AppState>) -> String {
 }
 // </setup>
 
-// <setup_mutable>
-struct AppStateWithCounter {
-    counter: Mutex<i32>, // <- Mutex is necessary to mutate safely across threads
-}
-
-async fn _index(data: web::Data<AppStateWithCounter>) -> String {
-    let mut counter = data.counter.lock().unwrap(); // <- get counter's MutexGuard
-    *counter += 1; // <- access counter inside MutexGuard
-
-    format!("Request number: {}", counter) // <- response with count
-}
-// </setup_mutable>
-
-// <make_app_mutable>
-#[actix_rt::main]
-async fn main() -> std::io::Result<()> {
-    let counter = web::Data::new(AppStateWithCounter {
-        counter: Mutex::new(0),
-    });
-
-    HttpServer::new(move || {
-        // move counter into the closure
-        App::new()
-            // Note: using app_data instead of data
-            .app_data(counter.clone()) // <- register the created data
-            .route("/", web::get().to(_index))
-    })
-    .bind("127.0.0.1:8088")?
-    .run()
-    .await
-}
-// </make_app_mutable>
-
 // <start_app>
 #[actix_rt::main]
 async fn main() -> std::io::Result<()> {
